[#24] feat : ChooseBank 화면 관련 Composable 분리
